Time: 0 ms (100%), Space: 41 MB (50.1%) - LeetHub

+++ 0119-pascals-triangle-ii/0119-pascals-triangle-ii.java
@@ -1,10 +1,10 @@
 class Solution {
     public List<Integer> getRow(int rowIndex) {
         List<Integer> list = new ArrayList<>();
-        int num = 1;
+        double num = 1;
         
         for (int i = 0; i <= rowIndex; i++) {
-            list.add(num);
+            list.add((int)num);
             num = num * (rowIndex - i) / (i + 1);
         }
